+++ tasks/week1（基础考核）/task3.1/random_data_pubsub/msg/RandomData.msg
@@ -1,0 +1,2 @@
+float64 data
+string message
