+++ tasks/week1（基础考核）/task3.2/math_service/srv/MathOps.srv
@@ -1,0 +1,6 @@
+float64 a
+float64 b
+---
+float64 sum
+float64 product
+bool equal
